fixed bug, that merged cells were not parsed
`elementNode` was the `sheetdata` node and was searched for the `mergeCell` node which is incorrect, because the `mergeCell` node is not contained in `sheetData` node but instead in the parent node called `worksheet`.
This is why merged cells were not parsed at all.

The comment "// Remove those cells which are present inside the" is not complete and the codeblock below it was never really executed, because the `spannedCells` was empty because of the bug. Now that it is not empty anymore because the correct node is searched for merged cells, the loop throws an error. While the columns are iterated, the code tries to delete entries, which can not happen while iterating. When changing to coe to not use an iterate to loop, it works that the cells are deleted. however this code deletes just every cell, that is not a combined cell, which makes a woorksheet empty, except for the combined cells.

However it is not even neccecary to remove any cells. That is why the codeblock is deleted. Parsing merged cells works afterwards just fine.